Added exception to wish list service

+++ src/Application/Services/Implementation/WishListService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.Entity.Core;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -17,13 +18,15 @@ namespace Application.Services.Implementation
     {
         private readonly IUserResolverService _userResolverService;
         private readonly IRepository<Wish> _wishRepository;
+        private readonly IRepository<Book> _bookRepository;
         private readonly IPaginationService _paginationService;
 
-        public WishListService(IUserResolverService userResolverService, IRepository<Wish> wishRepository, IPaginationService paginationService)
+        public WishListService(IUserResolverService userResolverService, IRepository<Wish> wishRepository, IPaginationService paginationService, IRepository<Book> bookRepository)
         {
             _userResolverService = userResolverService;
             _wishRepository = wishRepository;
             _paginationService = paginationService;
+            _bookRepository = bookRepository;
         }
 
         public async Task<PaginationDto<BookGetDto>> GetWishesOfCurrentUser(PageableParams pageableParams)
@@ -41,6 +44,17 @@ namespace Application.Services.Implementation
         public async Task AddWish(int bookId)
         {
             var currentUserId = _userResolverService.GetUserId();
+            var book = await _bookRepository.FindByIdAsync(bookId);
+
+            if (book == null)
+            {
+                throw new ObjectNotFoundException($"There is no book with id = {bookId} in database");
+            }
+
+            if (book.UserId == currentUserId)
+            {
+                throw new InvalidOperationException("User cannot add his book to wish list");
+            }
 
             var newWish = new Wish
             {
@@ -57,6 +71,10 @@ namespace Application.Services.Implementation
             var currentUserId = _userResolverService.GetUserId();
 
             var wishForRemoving = await _wishRepository.FindByIdAsync(currentUserId, bookId);
+            if (wishForRemoving == null)
+            {
+                throw new ObjectNotFoundException($"There is no book with id = {bookId} for current user");
+            }
 
             _wishRepository.Remove(wishForRemoving);
             await _wishRepository.SaveChangesAsync();
